+++ testsuite/tests/th/T5700a.hs
@@ -8,8 +8,8 @@ class C a where
 
 mkC :: Name -> Q [Dec]
 mkC n = return
-  [InstanceD Nothing [] (AppT (ConT ''C) (ConT n))
+  [InstanceD Nothing Nothing [] (AppT (ConT ''C) (ConT n))
     [ FunD 'inlinable [Clause [WildP] (NormalB (ConE '())) []],
-      PragmaD (InlineP 'inlinable Inline FunLike AllPhases)    
-    ] 
+      PragmaD (InlineP 'inlinable Inline FunLike AllPhases)
+    ]
   ]