+++ testsuite/tests/th/T5886a.hs
@@ -11,5 +11,5 @@ class C α where
   type AT α ∷ Type
 
 bang ∷ DecsQ
-bang = return [InstanceD Nothing [] (AppT (ConT ''C) (ConT ''Int))
+bang = return [InstanceD Nothing Nothing [] (AppT (ConT ''C) (ConT ''Int))
                 [TySynInstD (TySynEqn Nothing (AppT (ConT ''AT) (ConT ''Int)) (ConT ''Int))]]