+++ testsuite/tests/th/T7532a.hs
@@ -10,6 +10,6 @@ class C a where
 
 bang' :: DecsQ
 bang' = return [
-     InstanceD Nothing [] (AppT (ConT ''C) (ConT ''Int)) [
+     InstanceD Nothing Nothing [] (AppT (ConT ''C) (ConT ''Int)) [
          DataInstD [] Nothing (AppT (ConT ''D) (ConT ''Int)) Nothing [
              NormalC (mkName "T") []] []]]